requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
Pillow==9.3.0
pygame==2.1.3.dev8
requests==2.28.1
urllib3==1.26.13
